align with R1.6.1, remove unnecessary changes and fixed WAs

+++ openfold/utils/feats.py
@@ -107,16 +107,9 @@ def build_template_pair_feat(
     dgram = torch.sum(
         (tpb[..., None, :] - tpb[..., None, :, :]) ** 2, dim=-1, keepdim=True
     )
-    # habana limit change: linspace has bug, now it is fixed in R1.5.0
-    # lower = torch.linspace(min_bin, max_bin, no_bins, device=tpb.device) ** 2 ->
-    # lower = torch.linspace(min_bin, max_bin, no_bins, device="cpu").to(tpb.device) ** 2
-    # 
-    # mul not support bool
-    # dgram = ((dgram > lower) * (dgram < upper)).type(dgram.dtype) - >
-    # dgram = ((dgram > lower).type(dgram.dtype) * (dgram < upper).type(dgram.dtype)
     lower = torch.linspace(min_bin, max_bin, no_bins, device=tpb.device) ** 2
     upper = torch.cat([lower[1:], lower.new_tensor([inf])], dim=-1)
-    dgram = (dgram > lower).type(dgram.dtype) * (dgram < upper).type(dgram.dtype)
+    dgram = ((dgram > lower) * (dgram < upper)).type(dgram.dtype)
 
     to_concat = [dgram, template_mask_2d[..., None]]
